# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ERP System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      parameters:
        - name: departmentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── TASKS ───────────────────────────────────────────────────────────────────
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: projectId
          in: query
          schema:
            type: ["integer", "null"]
        - name: assignedTo
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: priority
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tasks/{id}/comments:
    get:
      operationId: listTaskComments
      tags: [tasks]
      summary: List comments on a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskComment"
    post:
      operationId: createTaskComment
      tags: [tasks]
      summary: Add a comment to a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskCommentBody"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskComment"

  # ─── NOTIFICATIONS ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        status:
          type: string
        createdBy:
          type: integer
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        taskCount:
          type: integer
        completedTaskCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, status, createdBy, taskCount, completedTaskCount, createdAt]
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
      required: [name]
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]

    # ─── TASKS ─────────────────────────────────────────────────────────────────
    Task:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: ["integer", "null"]
        projectName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        assignedTo:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        createdBy:
          type: integer
        priority:
          type: string
        status:
          type: string
        dueDate:
          type: ["string", "null"]
        commentCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, priority, status, createdBy, commentCount, createdAt, updatedAt]
    CreateTaskBody:
      type: object
      properties:
        projectId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        assignedTo:
          type: ["integer", "null"]
        priority:
          type: string
        dueDate:
          type: ["string", "null"]
      required: [title, priority]
    UpdateTaskBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        assignedTo:
          type: ["integer", "null"]
        priority:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
    TaskComment:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
        userId:
          type: integer
        userFullName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        body:
          type: string
        createdAt:
          type: string
      required: [id, taskId, userId, userFullName, body, createdAt]
    CreateTaskCommentBody:
      type: object
      properties:
        body:
          type: string
      required: [body]

    # ─── NOTIFICATIONS ─────────────────────────────────────────────────────────
...