# lib/api-zod/src/generated/types/updateTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ERP System API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateTaskBody {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  assignedTo?: number | null;
  /** @nullable */
  priority?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  dueDate?: string | null;
}